fix: cost values are reading and storing properly in db.

+++ backend/app/models/conversation_details_db.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Integer, ForeignKey
+from sqlalchemy import Column, String, Integer, Numeric, ForeignKey
 from app.config.database import Base
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
@@ -11,7 +11,7 @@ class ConversationDetails(Base):
     timestamp = Column(String, index=True)
     date = Column(String, index=True)
     cost_credits = Column(Integer, index=True)
-    cost_usd = Column(Integer, index=True)
+    cost_usd = Column(Numeric(12,6), index=True)
     call_duration_secs = Column(Integer, index=True)
     messages_count = Column(Integer, index=True)
