+++ backend/app/models/usage_db.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from app.config.database import Base
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
@@ -10,7 +10,7 @@ class Usage(Base):
     total_conversations = Column(Integer, index=True)
     total_messages = Column(Integer, index=True)
     total_credits = Column(Integer, index=True)
-    total_cost_usd = Column(Integer, index=True)
+    total_cost_usd = Column(Numeric(12,6), index=True)
     total_call_duration_secs = Column(Integer, index=True)
     
     user = relationship("User", back_populates="usage")
